Fade to specific LED in m2 model
Add support for fading to a specific LED with the m2 version. Its documented to work with the set immediate too, but that wasn't actually working for me so I just did it for the fade command. I just built out support for the two leds that are in the m2, but there is an m3 it looks like. I should figure out how many leds that one has